load jQuery independently and not as a dependency for premise WP js. see description
I had to do this because a theme had to change the version of jQuery.
In order to do so, the theme has to remove the normal version and
insert the new version that the theme wants. By doing this, when
removing the normal version since we have it as a dependency, premise
wp would not load. Simply loading an additional version of jQuery
creates conflicts. This was the best solution.